dao interfaces

+++ SpringApplication/src/main/java/com/revature/application/dao/EmployeeDao.java
@@ -4,7 +4,7 @@ import com.revature.application.dao.beans.*;
 
 public interface EmployeeDao {
 	public boolean create(Employee emp);
-	public Employee read(int employee_id);
+	public Employee read(long employee_id);
 	public Employee read(String username);
 	public boolean update(Employee emp);
 	public boolean delete(Employee emp);
